list comments out
