+++ client/coral-framework/constants/user.js
@@ -1,3 +1,6 @@
 export const SAVE_BIO_REQUEST = 'SAVE_BIO_REQUEST';
 export const SAVE_BIO_SUCCESS = 'SAVE_BIO_SUCCESS';
 export const SAVE_BIO_FAILURE = 'SAVE_BIO_FAILURE';
+export const REQUEST_COMMENTS_BY_USER = 'REQUEST_COMMENTS_BY_USER';
+export const RECEIVE_COMMENTS_BY_USER = 'RECEIVE_COMMENTS_BY_USER';
+export const FAILURE_COMMENTS_BY_USER = 'FAILURE_COMMENTS_BY_USER';
