+++ client/coral-framework/reducers/items.js
@@ -17,7 +17,6 @@ export default (state = initialState, action) => {
     return state.setIn([action.item_type, action.id, action.property], fromJS(action.value));
   case actions.APPEND_ITEM_ARRAY:
     return state.updateIn([action.item_type, action.id, action.property], (prop) => {
-      console.log(action, prop);
       if (action.add_to_front) {
         return prop ? prop.unshift(fromJS(action.value)) : fromJS([action.value]);
       } else {
